% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DL_second_derivative.R
\name{calc_DL_second_derivative}
\alias{calc_DL_second_derivative}
\title{Calculate second-order derivative of fitted DL curve}
\usage{
calc_DL_second_derivative(fhat, t_dense)
}
\arguments{
\item{fhat}{Numeric vector, fitted values on dense grid}

\item{t_dense}{Numeric vector, DOY corresponding to fhat}
}
\value{
Numeric vector, second derivative
}
\description{
Calculate second-order derivative of fitted DL curve
}
