% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dynamic_threshold_year.R
\name{calc_dynamic_threshold_year}
\alias{calc_dynamic_threshold_year}
\title{Calculate dynamic relative amplitude threshold for one year (TIMESAT method 1)}
\usage{
calc_dynamic_threshold_year(fit_year, threshold = 0.5)
}
\arguments{
\item{fit_year}{List, single year DL fit}

\item{threshold}{Fraction of amplitude (default 0.5)}
}
\value{
Numeric threshold for this year
}
\description{
Calculate dynamic relative amplitude threshold for one year (TIMESAT method 1)
}
