% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative_threshold.R
\name{calc_relative_threshold_pixel}
\alias{calc_relative_threshold_pixel}
\title{Calculate relative amplitude threshold for a pixel (Timesat method 3)}
\usage{
calc_relative_threshold_pixel(fits, threshold = 0.5)
}
\arguments{
\item{fits}{List of yearly DL fits}

\item{threshold}{Fraction of relative amplitude (default 0.5)}
}
\value{
Numeric, relative threshold for all years of this pixel
}
\description{
Calculate relative amplitude threshold for a pixel (Timesat method 3)
}
