% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pheno_relative_DL.R
\name{extract_pheno_DL_year}
\alias{extract_pheno_DL_year}
\title{Extract phenology (SOS/EOS/POS) from DL curve using threshold}
\usage{
extract_pheno_DL_year(fit_year, thr)
}
\arguments{
\item{fit_year}{List, single year DL fit}

\item{thr}{Numeric, threshold value}
}
\value{
Named vector c(SOS, EOS, POS)
}
\description{
Extract phenology (SOS/EOS/POS) from DL curve using threshold
}
