% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pheno_pixel_D2.R
\name{extract_pheno_pixel_D2}
\alias{extract_pheno_pixel_D2}
\title{Extract phenology for one pixel using second-derivative method}
\usage{
extract_pheno_pixel_D2(fits, years_sel = NULL, spring_end = 182)
}
\arguments{
\item{fits}{List of yearly DL fits}

\item{years_sel}{Optional integer vector}

\item{spring_end}{DOY defining first half year (default 182)}
}
\description{
Extract phenology for one pixel using second-derivative method
}
