% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pheno_pixel_TS1.R
\name{extract_pheno_pixel_TS1}
\alias{extract_pheno_pixel_TS1}
\title{Extract phenology for one pixel using TIMESAT method 1 (dynamic threshold)}
\usage{
extract_pheno_pixel_TS1(fits, threshold = 0.5, years_sel = NULL)
}
\arguments{
\item{fits}{List of yearly DL fits (output of fit_pixel_DL_yearly)}

\item{threshold}{Numeric, relative amplitude fraction}
}
\value{
Matrix with rows = years, cols = SOS/EOS/POS
}
\description{
Extract phenology for one pixel using TIMESAT method 1 (dynamic threshold)
}
