% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_first_local_max.R
\name{find_first_local_max}
\alias{find_first_local_max}
\title{Find first local maximum}
\usage{
find_first_local_max(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Index of first local maximum, or NA
}
\description{
Find first local maximum
}
