% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_pheno_extraction.R
\name{global_pheno_extraction}
\alias{global_pheno_extraction}
\title{Extract phenology for all pixels and years}
\usage{
global_pheno_extraction(
  r,
  doy,
  years,
  outdir,
  threshold = 0.5,
  years_sel = NULL
)
}
\arguments{
\item{r}{SpatRaster, raster stack (multi-layer GPP)}

\item{doy}{Numeric, DOY vector}

\item{years}{Integer, year vector}

\item{outdir}{Character, folder to save yearly single-band TIFFs}

\item{threshold}{Numeric, relative amplitude threshold fraction}

\item{years_sel}{Numeric, subset of years (optional)}
}
\description{
Extract phenology for all pixels and years
}
