% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_pheno_extraction_D2.R
\name{global_pheno_extraction_D2}
\alias{global_pheno_extraction_D2}
\title{Global phenology extraction using second-derivative (D2) method}
\usage{
global_pheno_extraction_D2(r, doy, years, outdir, years_sel = NULL, min_n = 10)
}
\arguments{
\item{r}{SpatRaster, NDVI time series}

\item{doy}{Numeric vector, day-of-year for each layer}

\item{years}{Integer vector, year for each layer}

\item{outdir}{Character, output directory}

\item{years_sel}{Optional integer vector, subset of years}

\item{min_n}{Integer, minimum valid observations per year}
}
\value{
A list with SOS, POS, EOS matrices (invisible)
}
\description{
Global phenology extraction using second-derivative (D2) method
}
