% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_pheno_extraction_TS1.R
\name{global_pheno_extraction_TS1}
\alias{global_pheno_extraction_TS1}
\title{Global phenology extraction using TIMESAT method 1 (dynamic threshold)}
\usage{
global_pheno_extraction_TS1(
  r,
  doy,
  years,
  outdir,
  threshold = 0.5,
  years_sel = NULL
)
}
\arguments{
\item{r}{SpatRaster, raster time series}

\item{doy}{Numeric vector, day of year}

\item{years}{Integer vector, year for each layer}

\item{outdir}{Character, output folder}

\item{threshold}{Numeric, relative amplitude fraction}

\item{years_sel}{Optional integer vector, subset of years}
}
\description{
Global phenology extraction using TIMESAT method 1 (dynamic threshold)
}
