% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raster_ts.R
\name{read_raster_ts}
\alias{read_raster_ts}
\title{Read raster time series from a folder}
\usage{
read_raster_ts(indir, pattern = "\\\\.tif$")
}
\arguments{
\item{indir}{Character, folder containing monthly/regular TIFFs}

\item{pattern}{Regex pattern to select files (default: "\\.tif$")}
}
\value{
A list containing SpatRaster and time vector
}
\description{
Read raster time series from a folder
}
