% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raster_ts_doy.R
\name{read_raster_ts_doy}
\alias{read_raster_ts_doy}
\title{Read raster time series with YYYY_DOY naming}
\usage{
read_raster_ts_doy(indir, pattern = "\\\\.tif$")
}
\arguments{
\item{indir}{Character, folder containing TIFFs named as *_YYYY_DDD.tif}

\item{pattern}{Regex pattern to select files}
}
\value{
A list with raster, dates, years, doy
}
\description{
Read raster time series with YYYY_DOY naming
}
